Actualizacion de DB por fallas, mejoras en back y front (skills)

+++ yopro/BackEnd/src/main/java/com/portfolio/mrn/Repository/ISkillRepository.java
@@ -1,14 +1,15 @@
 package com.portfolio.mrn.Repository;
 
 import com.portfolio.mrn.Entity.Skill;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
+
+
 @Repository
-public interface ISkillRepository extends JpaRepository<Skill, Long>{
+public interface ISkillRepository extends JpaRepository<Skill, Integer>{
 	public Optional<Skill> findByNombreSkill(String nombreSkill);
 	public boolean existsByNombreSkill(String nombreSkill);
 }
